VSS signal fixes
Changes:
- Fix sets of HVAC and volume control signals to update the
  actuator target value instead of the current value.  With the
  KUKSA.val ecosystem, it is the actuator implementation / backend
  that updates the current value after actuating.
- Add authorization token to subscribe request metadata if it is
  present.
- Switch type used for engine speed to double to match VSS signal
  definition.  This fixes receiving updates, and avoids needing
  to convert types as well.
- Added logic to push out the HVAC fan speed setting and handle
  external updates.
- Remove unused core/constants/paths.dart file to avoid confusion
  over VSS signals used and their variable naming.

Bug-AGL: SPEC-4999

+++ lib/data/models/vehicle.dart
@@ -12,7 +12,7 @@ class Vehicle {
   final int mediaVolume;
   final bool isChildLockActiveLeft;
   final bool isChildLockActiveRight;
-  final int engineSpeed;
+  final double engineSpeed;
   final int frontLeftTire;
   final int frontRightTire;
   final int rearLeftTire;
@@ -101,7 +101,7 @@ class Vehicle {
       int? mediaVolume,
       bool? isChildLockActiveLeft,
       bool? isChildLockActiveRight,
-      int? engineSpeed,
+      double? engineSpeed,
       int? frontLeftTire,
       int? frontRightTire,
       int? rearLeftTire,
